Refactor components and add documentation
